Fix print_progress type and API failure handling - use double for float values, silently ignore individual API failures

+++ bundles/org.openhab.binding.raise3d/src/main/java/org/openhab/binding/raise3d/internal/api/Raise3dPrinterData.java
@@ -35,7 +35,7 @@ public class Raise3dPrinterData {
     public @Nullable String jobStatus;
 
     @SerializedName("print_progress")
-    public int printProgress;
+    public double printProgress;
 
     @SerializedName("printed_layer")
     public int printedLayer;
